+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Allow description of transactions/splits to be 1024 characters long
 - Trim description before importing transactions
+- Enable app for NC 32
 
 ## [0.30.0] - 2025-03-31